Added SendPayload message, fix some indentation and change client id from Integer to Long

+++ src/main/java/com/company/common/messages/serverToCLI/ClientsAndStatuses.java
@@ -12,7 +12,7 @@ import java.util.Map;
 @Setter
 @NoArgsConstructor
 @ToString(callSuper = true)
-public class ClientsAndStatuses extends BaseServerToCLI{
+public class ClientsAndStatuses extends BaseServerToCLI {
 
-    private Map<Integer, ClientAndServerStatus> clientsAndStatuses;
+    private Map<Long, ClientAndServerStatus> clientsAndStatuses;
 }
